Move toArcaneSchema method to a new class
Minor refactoring
