+++ src/main/scala/services/mssql/query/LazyQueryResult.scala
@@ -1,8 +1,8 @@
 package com.sneaksanddata.arcane.framework
 package services.mssql.query
 
+import Utils.SqlUtils.toArcaneType
 import models.{DataCell, DataRow}
-import services.mssql.MsSqlConnection.toArcaneType
 import services.mssql.base.{CanPeekHead, QueryResult, ResultSetOwner}
 
 import java.sql.{ResultSet, Statement}
