+++ src/main/scala/services/mssql/query/ScalarQueryResult.scala
@@ -2,7 +2,6 @@ package com.sneaksanddata.arcane.framework
 package services.mssql.query
 
 import models.{DataCell, DataRow}
-import services.mssql.MsSqlConnection.toArcaneType
 import services.mssql.base.{QueryResult, ResultSetOwner}
 import services.mssql.query.LazyQueryResult
