+++ version.sbt
@@ -1,1 +1,1 @@
-ThisBuild / version := "0.1.1-SNAPSHOT"
+ThisBuild / version := "0.1.3-dev1-SNAPSHOT"
